StarlingX: Kernel Threads Workqueues IRQs
If the kthread_cpus boot arg is set it means we want to affine
kernel threads to the specified CPU mask as much as possible
in order to avoid doing work on other CPUs.

In this commit we extend the meaning of that boot arg to also
apply to the CPU affinity of unbound and ordered workqueues.

We also use the kthread_cpus value to determine the default irq
affinity.  Specifically, as long as the previously-calculated
irq affinity intersects with the kthread_cpus affinity then we'll
use the intersection of the two as the default irq affinity.

Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
[VT: replacing spaces with tabs.  Performed tests]
Signed-off-by: Vu Tran <vu.tran@windriver.com>
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>

Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>