StarlingX: Kernel Threads iSCSI
If the kthread_cpus boot arg is set it means we want to affine
kernel threads to the specified CPU mask as much as possible
in order to avoid doing work on other CPUs.

Do not allow them to run on other CPUs to prevent interference
with VMs.

In this commit we extend the meaning of that boot arg to also
apply to the affinity of iSCSI.

Signed-off-by: Alex Kozyrev <alex.kozyrev@windriver.com>
Signed-off-by: Jim Somerville <Jim.Somerville@windriver.com>

Signed-off-by: Abraham Arce <abraham.arce.moreno@intel.com>